# api_contract_basic.openapi.yaml
openapi: 3.0.3
info:
  title: 教学事务管理系统基础接口
  description: 基于当前基础接口契约整理的 OpenAPI 文档，用于导入 Apifox 调试。
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: 本地 Spring Boot 后端服务
...
paths:
...
  /api/courses:
    get:
      tags:
        - Common
      summary: 查询课程基础列表
      operationId: listCourses
      parameters:
        - $ref: "#/components/parameters/DeptIdQuery"
        - $ref: "#/components/parameters/KeywordQuery"
      responses:
        "200":
          description: 课程列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResultBase"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Course"
  /api/admin/students:
    get:
      tags:
        - Admin
      summary: 查询学生列表
      operationId: adminListStudents
      parameters:
        - $ref: "#/components/parameters/DeptIdQuery"
        - $ref: "#/components/parameters/KeywordQuery"
      responses:
        "200":
          description: 学生列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResultBase"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Student"
    post:
      tags:
        - Admin
      summary: 新增学生
      operationId: adminCreateStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentRequest"
      responses:
        "200":
          description: 新增成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResultBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Student"
        "400":
          $ref: "#/components/responses/ErrorResponse"
  /api/admin/students/{studentId}:
    put:
      tags:
        - Admin
      summary: 修改学生
      operationId: adminUpdateStudent
      parameters:
        - name: studentId
          in: path
          required: true
          description: 学生编号
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentRequest"
      responses:
        "200":
          description: 修改成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResultBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Student"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /api/admin/teachers:
    get:
      tags:
        - Admin
      summary: 查询教师列表
      operationId: adminListTeachers
      parameters:
        - $ref: "#/components/parameters/DeptIdQuery"
        - $ref: "#/components/parameters/KeywordQuery"
      responses:
        "200":
          description: 教师列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResultBase"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Teacher"
    post:
      tags:
        - Admin
      summary: 新增教师
      operationId: adminCreateTeacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherRequest"
      responses:
        "200":
          description: 新增成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResultBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Teacher"
        "400":
          $ref: "#/components/responses/ErrorResponse"
  /api/admin/teachers/{staffId}:
    put:
      tags:
        - Admin
      summary: 修改教师
      operationId: adminUpdateTeacher
      parameters:
        - name: staffId
          in: path
          required: true
          description: 教师编号
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherRequest"
      responses:
        "200":
          description: 修改成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResultBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Teacher"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /api/admin/courses:
    get:
      tags:
        - Admin
      summary: 查询课程列表
      operationId: adminListCourses
      parameters:
        - $ref: "#/components/parameters/DeptIdQuery"
        - $ref: "#/components/parameters/KeywordQuery"
      responses:
        "200":
          description: 课程列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResultBase"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Course"
    post:
      tags:
        - Admin
      summary: 新增课程
      operationId: adminCreateCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseRequest"
      responses:
        "200":
          description: 新增成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResultBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Course"
        "400":
          $ref: "#/components/responses/ErrorResponse"
  /api/admin/courses/{courseId}:
    put:
      tags:
        - Admin
      summary: 修改课程
      operationId: adminUpdateCourse
      parameters:
        - name: courseId
          in: path
          required: true
          description: 课程编号
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseRequest"
      responses:
        "200":
          description: 修改成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResultBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Course"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /api/admin/classes:
    get:
      tags:
        - Admin
      summary: 查询开课列表
      operationId: adminListClasses
      parameters:
        - $ref: "#/components/parameters/SemesterQuery"
        - $ref: "#/components/parameters/CourseIdQuery"
        - $ref: "#/components/parameters/StaffIdQuery"
      responses:
        "200":
          description: 开课列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResultBase"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ClassInfo"
    post:
      tags:
        - Admin
      summary: 新增开课
      operationId: adminCreateClass
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInfoRequest"
      responses:
        "200":
          description: 新增成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResultBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ClassInfo"
        "400":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  parameters:
...
    DeptIdQuery:
      name: deptId
      in: query
      required: false
      description: 院系编号
      schema:
        type: string
    KeywordQuery:
      name: keyword
      in: query
      required: false
      description: 模糊查询关键字
      schema:
        type: string
    SemesterQuery:
      name: semester
      in: query
      required: false
      description: 学期，例如 201202
      schema:
        type: string
...
    CourseIdQuery:
      name: courseId
      in: query
      required: false
      description: 课程编号
      schema:
        type: string
    StaffIdQuery:
      name: staffId
      in: query
      required: false
      description: 教师编号
      schema:
        type: string
...
  responses:
    ErrorResponse:
      description: 业务错误或系统错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResult"
  schemas:
...
    ResultBase:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          nullable: true
    ErrorResult:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: 参数错误
        data:
          nullable: true
          example: null
...
    Student:
      type: object
      properties:
        studentId:
          type: string
          example: "1102"
        name:
          type: string
          example: 刘晓明
        sex:
          type: string
          nullable: true
          enum:
            - 男
            - 女
          example: 男
        dateOfBirth:
          type: string
          format: date
          nullable: true
          example: "1992-12-08"
        nativePlace:
          type: string
          nullable: true
          example: 安徽
        mobilePhone:
          type: string
          nullable: true
          example: "18913457890"
        deptId:
          type: string
          nullable: true
          example: "01"
        status:
          type: string
          nullable: true
          example: 在读
    StudentRequest:
      allOf:
        - $ref: "#/components/schemas/Student"
      required:
        - studentId
        - name
    Teacher:
      type: object
      properties:
        staffId:
          type: string
          example: "0101"
        name:
          type: string
          example: 陈迪茂
        sex:
          type: string
          nullable: true
          enum:
            - 男
            - 女
          example: 男
        dateOfBirth:
          type: string
          format: date
          nullable: true
          example: "1983-03-06"
        professionalTitle:
          type: string
          nullable: true
          example: 副教授
        salary:
          type: number
          format: double
          nullable: true
          example: 7567.00
        deptId:
          type: string
          nullable: true
          example: "01"
    TeacherRequest:
      allOf:
        - $ref: "#/components/schemas/Teacher"
      required:
        - staffId
        - name
    Course:
      type: object
      properties:
        courseId:
          type: string
          example: "08305002"
        courseName:
          type: string
          example: 数据库原理
        credit:
          type: integer
          nullable: true
          example: 4
        creditHours:
          type: integer
          nullable: true
          example: 50
        deptId:
          type: string
          nullable: true
          example: "01"
    CourseRequest:
      allOf:
        - $ref: "#/components/schemas/Course"
      required:
        - courseId
        - courseName
    ClassInfo:
      type: object
      properties:
        semester:
          type: string
          example: "201202"
        courseId:
          type: string
          example: "08305002"
        staffId:
          type: string
          example: "0101"
        classTime:
          type: string
          nullable: true
          example: 星期三1-4
    ClassInfoRequest:
      allOf:
        - $ref: "#/components/schemas/ClassInfo"
      required:
        - semester
        - courseId
        - staffId
...